Custom draw popup menus in ListView control
Construct a menu containing the actions from the action list for the package
or category, and if one is selected, apply it.

This lets us remove packagemeta::set_action() which implements the strange
UX of cycling around actions without showing what the possibilities are.

This somewhat emulates a BS_SPLITBUTTON style control.  The 'popup' cell has
a visual hint that clicking on it opens a menu (a 'combox scrollbar'), and
the popup menu is located at the position of the click.

v2:
Factor out popup_menu, for future use by keyboard accelerators